fixup(l1): reconcile focil rebase against main
Align targetGasLimit with main's required-u64 (execution-apis#796): make
V5 target_gas_limit required to match V4, reject malformed V5 attributes in
parse_v5. Fix main API drift in FOCIL code (tx.hash(&NativeCrypto),
bal: Option<Arc<BlockAccessList>>, transaction_intrinsic_gas sender arg).
Relocate branch inline tests to test/tests/ per project convention.
