+++ crates/networking/rpc/types/fork_choice.rs
@@ -59,13 +59,12 @@ pub struct PayloadAttributesV5 {
     pub inclusion_list_transactions: Vec<Bytes>,
     // execution-apis#796: CL-supplied target gas limit, carried forward from
     // V4. V5 attributes are a superset of V4 (FOCIL only adds the IL), so the
-    // gas target stays available on the FOCIL local-build path. Required when
-    // Amsterdam/Hegotá is active — presence is enforced in
-    // `validate_attributes_v5`, matching `validate_attributes_v4`. Kept Option
-    // so an absent/null value yields a clean InvalidPayloadAttributes error
-    // rather than a serde parse failure (same pattern as V4).
-    #[serde(default, with = "serde_utils::u64::hex_str_opt")]
-    pub target_gas_limit: Option<u64>,
+    // gas target stays available on the FOCIL local-build path. Required (FOCIL
+    // runs on Hegotá, post-Amsterdam, where the gas target is mandatory); an
+    // absent field fails deserialization and the FCUv5 request is rejected,
+    // matching V4.
+    #[serde(with = "serde_utils::u64::hex_str")]
+    pub target_gas_limit: u64,
 }
 
 impl From<&PayloadAttributesV5> for PayloadAttributesV4 {
@@ -102,151 +101,5 @@ impl From<PayloadStatus> for ForkChoiceResponse {
             payload_status: value,
             payload_id: None,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn payload_attributes_v5_round_trips_with_inclusion_list() {
-        let json = r#"{
-            "timestamp": "0x6846fb2",
-            "prevRandao": "0x2971eefd1f71f3548728cad87c16cc91b979ef035054828c59a02e49ae300a84",
-            "suggestedFeeRecipient": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
-            "withdrawals": [],
-            "parentBeaconBlockRoot": "0x4029a2342bb6d54db91457bc8e442be22b3481df8edea24cc721f9d0649f65be",
-            "slotNumber": "0x10",
-            "inclusionListTransactions": ["0xdeadbeef", "0x01020304"]
-        }"#;
-
-        let attrs: PayloadAttributesV5 =
-            serde_json::from_str(json).expect("V5 attributes deserialize");
-
-        assert_eq!(attrs.timestamp, 0x6846fb2);
-        assert_eq!(attrs.slot_number, 0x10);
-        assert_eq!(
-            attrs.suggested_fee_recipient,
-            Address::from_slice(
-                &hex::decode("8943545177806ed17b9f23f0a21ee5948ecaa776")
-                    .expect("decode fee recipient")
-            )
-        );
-        assert!(attrs.withdrawals.is_some());
-        assert!(attrs.parent_beacon_block_root.is_some());
-        assert_eq!(attrs.inclusion_list_transactions.len(), 2);
-        assert_eq!(
-            attrs.inclusion_list_transactions[0].as_ref(),
-            &[0xde, 0xad, 0xbe, 0xef][..]
-        );
-        assert_eq!(
-            attrs.inclusion_list_transactions[1].as_ref(),
-            &[0x01, 0x02, 0x03, 0x04][..]
-        );
-
-        let serialized = serde_json::to_string(&attrs).expect("V5 attributes serialize");
-        assert!(
-            serialized.contains("\"inclusionListTransactions\":[\"0xdeadbeef\",\"0x01020304\"]")
-        );
-        assert!(serialized.contains("\"slotNumber\":\"0x10\""));
-
-        let reparsed: PayloadAttributesV5 =
-            serde_json::from_str(&serialized).expect("V5 attributes round-trip");
-        assert_eq!(
-            reparsed.inclusion_list_transactions,
-            attrs.inclusion_list_transactions
-        );
-        assert_eq!(reparsed.timestamp, attrs.timestamp);
-        assert_eq!(reparsed.slot_number, attrs.slot_number);
-    }
-
-    #[test]
-    fn payload_attributes_v5_accepts_empty_inclusion_list() {
-        let json = r#"{
-            "timestamp": "0x6846fb2",
-            "prevRandao": "0x2971eefd1f71f3548728cad87c16cc91b979ef035054828c59a02e49ae300a84",
-            "suggestedFeeRecipient": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
-            "withdrawals": [],
-            "parentBeaconBlockRoot": "0x4029a2342bb6d54db91457bc8e442be22b3481df8edea24cc721f9d0649f65be",
-            "slotNumber": "0x10",
-            "inclusionListTransactions": []
-        }"#;
-
-        let attrs: PayloadAttributesV5 =
-            serde_json::from_str(json).expect("V5 attributes deserialize");
-        assert!(attrs.inclusion_list_transactions.is_empty());
-
-        let serialized = serde_json::to_string(&attrs).expect("V5 attributes serialize");
-        assert!(serialized.contains("\"inclusionListTransactions\":[]"));
-    }
-
-    #[test]
-    fn payload_attributes_v5_to_v4_propagates_gas_limit_and_drops_il() {
-        let attrs_v5 = PayloadAttributesV5 {
-            timestamp: 0x6846fb2,
-            prev_randao: H256::zero(),
-            suggested_fee_recipient: Address::zero(),
-            withdrawals: Some(vec![]),
-            parent_beacon_block_root: Some(H256::zero()),
-            slot_number: 0x10,
-            inclusion_list_transactions: vec![Bytes::from_static(&[0xde, 0xad, 0xbe, 0xef])],
-            target_gas_limit: Some(50_000_000),
-        };
-
-        let attrs_v4: PayloadAttributesV4 = (&attrs_v5).into();
-        assert_eq!(attrs_v4.timestamp, attrs_v5.timestamp);
-        assert_eq!(attrs_v4.slot_number, attrs_v5.slot_number);
-        assert_eq!(attrs_v4.prev_randao, attrs_v5.prev_randao);
-        assert_eq!(
-            attrs_v4.suggested_fee_recipient,
-            attrs_v5.suggested_fee_recipient
-        );
-        assert_eq!(attrs_v4.withdrawals, attrs_v5.withdrawals);
-        assert_eq!(
-            attrs_v4.parent_beacon_block_root,
-            attrs_v5.parent_beacon_block_root
-        );
-        // execution-apis#796: the gas target must survive the V5->V4 downgrade.
-        assert_eq!(attrs_v4.target_gas_limit, attrs_v5.target_gas_limit);
-    }
-
-    #[test]
-    fn payload_attributes_v5_parses_target_gas_limit() {
-        let base = |extra: &str| {
-            format!(
-                r#"{{
-                    "timestamp": "0x65",
-                    "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000001",
-                    "suggestedFeeRecipient": "0x0000000000000000000000000000000000000002",
-                    "withdrawals": [],
-                    "parentBeaconBlockRoot": "0x0000000000000000000000000000000000000000000000000000000000000003",
-                    "slotNumber": "0x10",
-                    "inclusionListTransactions": []{extra}
-                }}"#
-            )
-        };
-
-        // present -> Some
-        let attrs: PayloadAttributesV5 =
-            serde_json::from_str(&base(r#", "targetGasLimit": "0x2faf080""#)).unwrap();
-        assert_eq!(attrs.target_gas_limit, Some(50_000_000));
-
-        // absent -> None (enforcement happens in validate_attributes_v5)
-        let attrs: PayloadAttributesV5 = serde_json::from_str(&base("")).unwrap();
-        assert!(attrs.target_gas_limit.is_none());
-
-        // explicit null -> None
-        let attrs: PayloadAttributesV5 =
-            serde_json::from_str(&base(r#", "targetGasLimit": null"#)).unwrap();
-        assert!(attrs.target_gas_limit.is_none());
-    }
-
-    #[test]
-    fn payload_attributes_v5_default_constructible() {
-        let attrs = PayloadAttributesV5::default();
-        assert_eq!(attrs.timestamp, 0);
-        assert_eq!(attrs.slot_number, 0);
-        assert!(attrs.inclusion_list_transactions.is_empty());
     }
 }
