+++ test/tests/blockchain/mod.rs
@@ -7,6 +7,8 @@ mod eip7702_revert_authority_tests;
 mod eip7702_zero_transfer_tests;
 mod explicit_blob_tx_tests;
 mod focil_tests;
+mod inclusion_list_builder_tests;
+mod inclusion_list_validator_tests;
 mod l1_tx_type_tests;
 mod logs_bloom_tests;
 mod mempool_tests;
