+++ test/tests/common/mod.rs
@@ -6,6 +6,7 @@ mod blobs_bundle_tests;
 mod code_serde_tests;
 mod eip7702_authorization_tests;
 mod frame_tx_validation_tests;
+mod inclusion_list_tests;
 mod legacy_signature_tests;
 mod logs_bloom_validation_tests;
 mod requests_eip8282_tests;
